Narrow the last on-disk field writes explicitly
The eight remaining -Wconversion warnings are value-narrowings, not sign
changes: mode/fsize into 16-bit on-disk slots, dirent inode numbers, and
the fsck -N/-C inode parsed from the command line.  Cast them to the field
width they store -- the on-disk fields really are 16 bits.

+++ fsck.filsys.c
@@ -123,11 +123,11 @@ int main(int argc, char **argv)
             break;
         case 'N':
             ncheck = 1;
-            ino = strtoul(optarg, NULL, 0);
+            ino = (uint32_t)strtoul(optarg, NULL, 0);
             break;
         case 'C':
             clri = 1;
-            ino = strtoul(optarg, NULL, 0);
+            ino = (uint32_t)strtoul(optarg, NULL, 0);
             break;
         default:
             fprintf(stderr,
